Move method getDefaultValidationRulesForFieldType to new model class
Moves PublicationFieldService::getDefaultValidationRulesForFieldType() to PublicationFieldValue::getDefaultValidationRulesForFieldType()
